Migrate atoms examples to the playground
