--- packages/docs/components/atoms/Button/stories/primary/story.md
@@ -1,5 +1,0 @@
----
-layout: none
----
-
-<RenderTwig :tpl-importer="() => import('./app.twig?raw')" />
