--- packages/docs/components/atoms/Button/stories/secondary-rounded/story.md
@@ -1,5 +1,0 @@
----
-layout: none
----
-
-<RenderTwig :tpl-importer="() => import('./app.twig?raw')" />
