+++ packages/docs/components/atoms/Cursor/examples.md
@@ -6,21 +6,7 @@ title: Cursor examples
 
 ## Simple
 
-<PreviewIframe src="./stories/story.html" />
-
-:::details Code
-
-<SimpleTabs :items="['app.twig', 'app.js']">
-  <template #content-1>
-
-<<< ./stories/app.twig
-
-  </template>
-  <template #content-2>
-
-<<< ./stories/app.js
-
-  </template>
-</SimpleTabs>
-
-:::
+<PreviewPlayground
+  :html="() => import('./stories/app.twig')"
+  :script="() => import('./stories/app.js?raw')"
+  />
