+++ packages/docs/components/atoms/Figure/examples.md
@@ -6,21 +6,7 @@ title: Figure examples
 
 ## Lazyload
 
-<PreviewIframe src="./stories/lazyload/story.html" />
-
-:::details Code
-
-<SimpleTabs :items="['app.twig', 'app.js']">
-  <template #content-1>
-
-<<< ./stories/lazyload/app.twig
-
-  </template>
-  <template #content-2>
-
-<<< ./stories/lazyload/app.js
-
-  </template>
-</SimpleTabs>
-
-:::
+<PreviewPlayground
+  :html="() => import('./stories/lazyload/app.twig')"
+  :script="() => import('./stories/lazyload/app.js?raw')"
+  />
