--- packages/docs/components/atoms/FigureTwicpics/stories/twicpics/Figure.js
@@ -1,11 +1,0 @@
-import { FigureTwicpics } from '@studiometa/ui';
-
-export default class Figure extends FigureTwicpics {
-  get domain() {
-    return 'studiometa.twic.pics';
-  }
-
-  get path() {
-    return 'ui';
-  }
-}
