+++ packages/docs/components/atoms/FigureTwicpics/stories/twicpics/app.js
@@ -1,5 +1,15 @@
 import { Base, createApp } from '@studiometa/js-toolkit';
-import Figure from './Figure.js';
+import { FigureTwicpics } from '@studiometa/ui';
+
+class Figure extends FigureTwicpics {
+  get domain() {
+    return 'studiometa.twic.pics';
+  }
+
+  get path() {
+    return 'ui';
+  }
+}
 
 class App extends Base {
   static config = {
