--- packages/docs/components/atoms/FigureTwicpics/stories/twicpics/story.md
@@ -1,6 +1,0 @@
----
-layout: none
----
-
-<RenderTwig :js-importer="() => import('./app.js')" :tpl-importer="() => import('./app.twig?raw')" />
-
