+++ packages/docs/components/atoms/IconInlineImg/examples.md
@@ -6,16 +6,6 @@ title: IconInlineImg examples
 
 ## Simple
 
-<PreviewIframe src="./stories/story.html" height="300px" />
-
-:::details Code
-
-<SimpleTabs :items="['app.twig']">
-  <template #content-1>
-
-<<< ./stories/app.twig
-
-  </template>
-</SimpleTabs>
-
-:::
+<PreviewPlayground
+  :html="() => import('./stories/app.twig')"
+  />
