+++ packages/docs/components/atoms/LargeText/examples.md
@@ -6,21 +6,8 @@ title: LargeText examples
 
 ## Simple
 
-<PreviewIframe src="./stories/story.html" />
+<PreviewPlayground
+  :html="() => import('./stories/app.twig')"
+  :script="() => import('./stories/app.js?raw')"
+  />
 
-:::details Code
-
-<SimpleTabs :items="['app.twig', 'app.js']">
-  <template #content-1>
-
-<<< ./stories/app.twig
-
-  </template>
-  <template #content-2>
-
-<<< ./stories/app.js
-
-  </template>
-</SimpleTabs>
-
-:::
