+++ packages/docs/components/atoms/LargeText/stories/app.twig
@@ -1,4 +1,4 @@
-<div class="h-96"></div>
+<div class="h-48"></div>
 
 {% include '@ui/atoms/LargeText/LargeText.twig' with {
   content: 'Lorem ipsum dolor sit amet elit.'
@@ -8,7 +8,7 @@
   attr: { data_option_sensitivity: - 1 }
 } %}
 
-<div class="h-96"></div>
+<div class="h-48"></div>
 
 {% include '@ui/atoms/LargeText/LargeText.twig' with {
   content: 'Hello world !',
