--- packages/docs/components/atoms/ScrollAnimation/stories/parallax-parent/ParallaxChild.js
@@ -1,15 +1,0 @@
-import { Figure, ScrollAnimationChild } from '@studiometa/ui';
-
-export default class ParallaxChild extends ScrollAnimationChild {
-  static config = {
-    ...ScrollAnimationChild.config,
-    name: 'ParallaxChild',
-    components: {
-      Figure,
-    },
-  };
-
-  get target() {
-    return this.$children.Figure[0].$el;
-  }
-}
