--- packages/docs/components/atoms/ScrollAnimation/stories/parallax/story.md
@@ -1,13 +1,0 @@
----
-layout: none
----
-
-<script setup>
-  import { onMounted } from 'vue';
-
-  onMounted(() => {
-    document.body.style.padding = '0';
-  });
-</script>
-
-<RenderTwig :js-importer="() => import('./app.js')" :tpl-importer="() => import('./app.twig?raw')" />
