--- packages/docs/components/atoms/ScrollAnimation/stories/parent/ScrollAnimationParent.js
@@ -1,13 +1,0 @@
-import {
-  ScrollAnimationParent as ScrollAnimationParentCore,
-  ScrollAnimationChild,
-} from '@studiometa/ui';
-
-export default class ScrollAnimationParent extends ScrollAnimationParentCore {
-  static config = {
-    name: 'ScrollAnimationParent',
-    components: {
-      ScrollAnimationChild,
-    },
-  };
-}
