--- packages/docs/components/atoms/ScrollReveal/stories/no-target-ref/story.md
@@ -1,5 +1,0 @@
----
-layout: none
----
-
-<RenderTwig :js-importer="() => import('./app.js')" :tpl-importer="() => import('./app.twig?raw')" />
